+++ src/comp124graphics/CanvasWindow.java
@@ -21,6 +21,11 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 
+import comp124graphics.events.MouseButtonEvent;
+import comp124graphics.events.MouseButtonEventHandler;
+import comp124graphics.events.MouseMotionEvent;
+import comp124graphics.events.MouseMotionEventHandler;
+
 /**
  * Creates a window frame and canvas panel that is used to draw graphical objects.
  * Created by bjackson on 9/13/2016.
