+++ src/comp124graphics/events/MouseButtonEvent.java
@@ -1,11 +1,11 @@
-package comp124graphics;
+package comp124graphics.events;
 
 import comp124graphics.Point;
 
 public final class MouseButtonEvent {
     private final Point position;
 
-    MouseButtonEvent(java.awt.event.MouseEvent sourceEvent) {
+    public MouseButtonEvent(java.awt.event.MouseEvent sourceEvent) {
         position = new Point(sourceEvent.getPoint());
     }
