+++ src/comp124graphics/events/MouseMotionEventHandler.java
@@ -1,4 +1,4 @@
-package comp124graphics;
+package comp124graphics.events;
 
 public interface MouseMotionEventHandler {
     void handleEvent(MouseMotionEvent event);
